restructure frontend folder
