+++ src/Umbraco.Commerce.Checkout/Client/package.json
@@ -22,6 +22,7 @@
     "typescript": "^5.4.5",
     "typescript-eslint": "^7.13.1",
     "vite": "^5.3.1",
+    "vite-plugin-static-copy": "^1.0.6",
     "vite-tsconfig-paths": "^4.3.2"
   },
   "author": "",
