+++ src/Umbraco.Commerce.Checkout/Client/src/backoffice/apis/install.api.ts
@@ -6,15 +6,14 @@ export type InstallUccApiResponse = {
 export const installUmbracoCommerceCheckoutAsync: (siteRootNodeId: string) => Promise<InstallUccApiResponse> = (siteRootNodeId: string) => {
     const response = fetch('/umbraco/management/api/v1/umbraco-commerce-checkout/install?siteRootNodeId=' + siteRootNodeId, {
         credentials: 'include',
-    })
-        .then(
-            (response: Response) => {
-                return response.json();
-            },
-            (reason: unknown) => ({
-                success: false,
-                message: JSON.stringify(reason),
-            })) as Promise<InstallUccApiResponse>;
+    }).then(
+        (response: Response) => {
+            return response.json() as Promise<InstallUccApiResponse>;
+        },
+        (reason: unknown) => ({
+            success: false,
+            message: JSON.stringify(reason),
+        })) as Promise<InstallUccApiResponse>;
 
     return response;
 };
