+++ src/Umbraco.Commerce.Checkout/Client/src/backoffice/modals/installer-modal.element.ts
@@ -5,7 +5,7 @@ import { ManifestModal, UmbModalExtensionElement } from '@umbraco-cms/backoffice
 import { UccInstallerModalSubmitValue } from './installer-modal.token';
 import { UmbInputDocumentElement } from '@umbraco-cms/backoffice/document';
 import type { UUIButtonState } from '@umbraco-cms/backoffice/external/uui';
-import { installUmbracoCommerceCheckoutAsync } from './apis';
+import { installUmbracoCommerceCheckoutAsync } from '../apis/install.api';
 import { UMB_NOTIFICATION_CONTEXT, UmbNotificationContext } from '@umbraco-cms/backoffice/notification';
