+++ src/Umbraco.Commerce.Checkout/Client/src/backoffice/modals/installer-modal.token.ts
