+++ src/Umbraco.Commerce.Checkout/Views/UmbracoCommerceCheckout/UmbracoCommerceCheckoutLayout.cshtml
@@ -29,7 +29,7 @@
     <meta name="description" content="" />
     <meta name="keywords" content="" />
     <meta name="robots" content="noindex, nofollow" />
-    <link href="/App_Plugins/UmbracoCommerceCheckout/dist/uccheckout.surface.css" rel="stylesheet">
+    <link href="/App_Plugins/UmbracoCommerceCheckout/uccheckout.surface.css" rel="stylesheet">
     @RenderSection("head", false)
 </head>
 <body>
@@ -228,7 +228,7 @@
         </symbol>
     </svg>
 
-    <script src="/App_Plugins/UmbracoCommerceCheckout/dist/uccheckout.surface.js" defer></script>
+    <script src="/App_Plugins/UmbracoCommerceCheckout/uccheckout.surface.js" defer async></script>
 
     @RenderSection("foot", false)
